# packages/contracts/openapi/amanor-v1.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "minLength": 1,
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "type": "string"
          },
          "requestId": {
            "minLength": 1,
            "type": "string"
          },
          "statusCode": {
            "maximum": 599,
            "minimum": 400,
            "type": "integer"
          },
          "timestamp": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "requestId",
          "timestamp"
        ],
        "type": "object"
      },
...
      "HardwareKeyAuthenticationDto": {
        "properties": {
          "ceremonyId": {
            "type": "string"
          },
          "response": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "ceremonyId",
          "response"
        ],
        "type": "object"
      },
      "HardwareKeyRegistrationDto": {
        "properties": {
          "ceremonyId": {
            "type": "string"
          },
          "label": {
            "maxLength": 80,
            "minLength": 1,
            "type": "string"
          },
          "response": {
            "additionalProperties": true,
            "type": "object"
          }
        },
        "required": [
          "ceremonyId",
          "response",
          "label"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Backend contracts for the public platform, CMS and operations console.",
    "title": "Project AMANOR API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
    "/v1/auth/audit": {
      "get": {
        "operationId": "AuthController_authenticationAudit",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Redacted, cursor-paginated authentication security events."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/audit/integrity": {
      "get": {
        "operationId": "AuthController_authenticationAuditIntegrity",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Authentication audit hash-chain integrity status."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/hardware-keys": {
      "get": {
        "operationId": "AuthController_hardwareKeyList",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/hardware-keys/{credentialId}": {
      "delete": {
        "operationId": "AuthController_hardwareKeyRevoke",
        "parameters": [
          {
            "in": "path",
            "name": "credentialId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/hardware-keys/authentication/options": {
      "post": {
        "operationId": "AuthController_hardwareKeyAuthenticationOptions",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/hardware-keys/authentication/verify": {
      "post": {
        "operationId": "AuthController_hardwareKeyAuthenticationVerify",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareKeyAuthenticationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/hardware-keys/registration/options": {
      "post": {
        "operationId": "AuthController_hardwareKeyRegistrationOptions",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/hardware-keys/registration/verify": {
      "post": {
        "operationId": "AuthController_hardwareKeyRegistrationVerify",
        "parameters": [
          {
            "in": "header",
            "name": "origin",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HardwareKeyRegistrationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            },
            "description": ""
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/invitations": {
      "post": {
        "operationId": "AuthController_invite",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdministratorInvitationDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "One-time administrator invitation created."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/invitations/accept": {
      "post": {
        "operationId": "AuthController_acceptInvitation",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationAcceptanceDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationAcceptanceResponseDto"
                }
              }
            },
            "description": "Password and MFA enrollment accepted; recovery codes shown once."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/invitations/setup": {
      "post": {
        "operationId": "AuthController_invitationSetup",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationSetupDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Valid invitation MFA enrollment details."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
    "/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Credentials verified and MFA challenge issued, or MFA verified and session created."
          },
          "401": {
            "description": "Credentials or MFA verification failed."
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error response. The status code is repeated in the bounded JSON envelope.",
            "headers": {
              "X-Request-ID": {
                "description": "Correlation identifier safe to quote to support.",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "tags": [
          "authentication"
        ]
      }
    },
...
  },
  "servers": [],
...
}